Split case editor module into parts
